Add another space to the playlist duration. Clean up the code.

+++ app/src/main/java/com/kabouzeid/gramophone/helper/MusicPlayerRemote.java
@@ -239,7 +239,7 @@ public class MusicPlayerRemote {
         return -1;
     }
 
-    public static int getQueueDurationMillis(int position) {
+    public static long getQueueDurationMillis(int position) {
         if (musicService != null) {
             return musicService.getQueueDurationMillis(position);
         }
